refactor: MemberInfoDTO에 조회한 카페 목록을 추가한다
