+++ backend/src/main/java/com/cafein/backend/api/member/service/MemberInfoService.java
@@ -22,7 +22,6 @@ public class MemberInfoService {
 
 	@Transactional(readOnly = true)
 	public MemberInfoResponseDTO getMemberInfo(final Long memberId) {
-
 		Member member = memberService.findMemberByMemberId(memberId);
 		List<Long> viewdCafeIds = viewedCafeService.findViewedCafes(memberId);
 		return MemberInfoResponseDTO.of(member, viewdCafeIds);
