+++ backend/src/main/java/com/cafein/backend/domain/viewedcafe/service/ViewedCafeService.java
@@ -1,11 +1,10 @@
-package com.cafein.backend.domain.viewedcafe;
+package com.cafein.backend.domain.viewedcafe.service;
 
 import java.util.List;
 
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import com.cafein.backend.domain.viewedcafe.entity.ViewedCafe;
 import com.cafein.backend.domain.viewedcafe.repository.ViewedCafeRepository;
 
 import lombok.RequiredArgsConstructor;
